+++ instrumentation/opentelemetry-instrumentation-redis/tests/test_redis.py
@@ -35,6 +35,15 @@ from opentelemetry.trace import SpanKind
 
 
 class TestRedis(TestBase):
+    def assert_span_count(self, count: int):
+        """
+        Assert that the memory exporter has the expected number of spans.
+        Returns the spans list if assertion passes
+        """
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), count)
+        return spans
+
     def setUp(self):
         super().setUp()
         RedisInstrumentor().instrument(tracer_provider=self.tracer_provider)
@@ -387,32 +396,216 @@ class TestRedis(TestBase):
 
 
 class TestRedisAsync(TestBase, IsolatedAsyncioTestCase):
+    def assert_span_count(self, count: int):
+        """
+        Assert that the memory exporter has the expected number of spans.
+        Returns the spans list if assertion passes
+        """
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), count)
+        return spans
+
     def setUp(self):
         super().setUp()
-        RedisInstrumentor().instrument(tracer_provider=self.tracer_provider)
+        self.instrumentor = RedisInstrumentor()
+        self.client: FakeRedis = FakeRedis()
 
-    def tearDown(self):
-        super().tearDown()
-        RedisInstrumentor().uninstrument()
+    @staticmethod
+    async def _redis_pipeline_operations(client: FakeRedis):
+        with pytest.raises(WatchError):
+            async with client.pipeline(transaction=False) as pipe:
+                await pipe.watch("a")
+                await client.set("a", "bad")
+                pipe.multi()
+                await pipe.set("a", "1")
+                await pipe.execute()
 
     @pytest.mark.asyncio
     async def test_watch_error_async(self):
-        async def redis_operations():
-            with pytest.raises(WatchError):
-                redis_client = FakeRedis()
-                async with redis_client.pipeline(transaction=False) as pipe:
-                    await pipe.watch("a")
-                    await redis_client.set("a", "bad")
-                    pipe.multi()
-                    await pipe.set("a", "1")
-                    await pipe.execute()
+        # this tests also ensures the response_hook is called
+        response_attr = "my.response.attribute"
+        count = 0
 
-        await redis_operations()
+        def response_hook(span, conn, args):
+            nonlocal count
+            if span and span.is_recording():
+                span.set_attribute(response_attr, count)
+                count += 1
+
+        self.instrumentor.instrument(
+            tracer_provider=self.tracer_provider, response_hook=response_hook
+        )
+        redis_client = FakeRedis()
+        await self._redis_pipeline_operations(redis_client)
+
+        # there should be 3 tests, we start watch operation and have 2 set operation on same key
+        spans = self.assert_span_count(3)
+
+        self.assertEqual(spans[0].attributes.get("db.statement"), "WATCH ?")
+        self.assertEqual(spans[0].kind, SpanKind.CLIENT)
+        self.assertEqual(spans[0].status.status_code, trace.StatusCode.UNSET)
+        self.assertEqual(spans[0].attributes.get(response_attr), 0)
+
+        for span_index, span in enumerate(spans[1:], 1):
+            self.assertEqual(span.attributes.get("db.statement"), "SET ? ?")
+            self.assertEqual(span.kind, SpanKind.CLIENT)
+            self.assertEqual(span.status.status_code, trace.StatusCode.UNSET)
+            self.assertEqual(span.attributes.get(response_attr), span_index)
+        RedisInstrumentor().uninstrument()
+
+    @pytest.mark.asyncio
+    async def test_watch_error_async_only_client(self):
+        self.instrumentor.instrument_client(
+            tracer_provider=self.tracer_provider, client=self.client
+        )
+        redis_client = FakeRedis()
+        await self._redis_pipeline_operations(redis_client)
+
+        spans = self.memory_exporter.get_finished_spans()
+
+        # there should be 3 tests, we start watch operation and have 2 set operation on same key
+        self.assertEqual(len(spans), 0)
+
+        # now with the instrumented client we should get proper spans
+        await self._redis_pipeline_operations(self.client)
 
         spans = self.memory_exporter.get_finished_spans()
 
         # there should be 3 tests, we start watch operation and have 2 set operation on same key
         self.assertEqual(len(spans), 3)
+
+        self.assertEqual(spans[0].attributes.get("db.statement"), "WATCH ?")
+        self.assertEqual(spans[0].kind, SpanKind.CLIENT)
+        self.assertEqual(spans[0].status.status_code, trace.StatusCode.UNSET)
+
+        for span in spans[1:]:
+            self.assertEqual(span.attributes.get("db.statement"), "SET ? ?")
+            self.assertEqual(span.kind, SpanKind.CLIENT)
+            self.assertEqual(span.status.status_code, trace.StatusCode.UNSET)
+        RedisInstrumentor().uninstrument_client(self.client)
+
+    @pytest.mark.asyncio
+    async def test_request_response_hooks(self):
+        request_attr = "my.request.attribute"
+        response_attr = "my.response.attribute"
+
+        def request_hook(span, conn, args, kwargs):
+            if span and span.is_recording():
+                span.set_attribute(request_attr, args[0])
+
+        def response_hook(span, conn, args):
+            if span and span.is_recording():
+                span.set_attribute(response_attr, args)
+
+        self.instrumentor.instrument(
+            tracer_provider=self.tracer_provider,
+            request_hook=request_hook,
+            response_hook=response_hook,
+        )
+        await self.client.set("key", "value")
+
+        spans = self.assert_span_count(1)
+
+        span = spans[0]
+        self.assertEqual(span.attributes.get(request_attr), "SET")
+        self.assertEqual(span.attributes.get(response_attr), True)
+        self.instrumentor.uninstrument()
+
+    @pytest.mark.asyncio
+    async def test_request_response_hooks_connection_only(self):
+        request_attr = "my.request.attribute"
+        response_attr = "my.response.attribute"
+
+        def request_hook(span, conn, args, kwargs):
+            if span and span.is_recording():
+                span.set_attribute(request_attr, args[0])
+
+        def response_hook(span, conn, args):
+            if span and span.is_recording():
+                span.set_attribute(response_attr, args)
+
+        self.instrumentor.instrument_client(
+            client=self.client,
+            tracer_provider=self.tracer_provider,
+            request_hook=request_hook,
+            response_hook=response_hook,
+        )
+        await self.client.set("key", "value")
+
+        spans = self.assert_span_count(1)
+
+        span = spans[0]
+        self.assertEqual(span.attributes.get(request_attr), "SET")
+        self.assertEqual(span.attributes.get(response_attr), True)
+        # fresh client should not record any spans
+        fresh_client = FakeRedis()
+        self.memory_exporter.clear()
+        await fresh_client.set("key", "value")
+        self.assert_span_count(0)
+        self.instrumentor.uninstrument_client(self.client)
+        # after un-instrumenting the query should not be recorder
+        await self.client.set("key", "value")
+        spans = self.assert_span_count(0)
+
+
+class TestRedisInstance(TestBase):
+    def assert_span_count(self, count: int):
+        """
+        Assert that the memory exporter has the expected number of spans.
+        Returns the spans list if assertion passes
+        """
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), count)
+        return spans
+
+    def setUp(self):
+        super().setUp()
+        self.client = fakeredis.FakeStrictRedis()
+        RedisInstrumentor().instrument_client(
+            client=self.client, tracer_provider=self.tracer_provider
+        )
+
+    def tearDown(self):
+        super().tearDown()
+        RedisInstrumentor().uninstrument_client(self.client)
+
+    def test_only_client_instrumented(self):
+        redis_client = redis.Redis()
+
+        with mock.patch.object(redis_client, "connection"):
+            redis_client.get("key")
+
+        spans = self.assert_span_count(0)
+
+        # now use the test client
+        with mock.patch.object(self.client, "connection"):
+            self.client.get("key")
+        spans = self.assert_span_count(1)
+        span = spans[0]
+        self.assertEqual(span.name, "GET")
+        self.assertEqual(span.kind, SpanKind.CLIENT)
+
+    @staticmethod
+    def redis_operations(client):
+        with pytest.raises(WatchError):
+            pipe = client.pipeline(transaction=True)
+            pipe.watch("a")
+            client.set("a", "bad")  # This will cause the WatchError
+            pipe.multi()
+            pipe.set("a", "1")
+            pipe.execute()
+
+    def test_watch_error_sync_only_client(self):
+        redis_client = fakeredis.FakeStrictRedis()
+
+        self.redis_operations(redis_client)
+
+        self.assert_span_count(0)
+
+        self.redis_operations(self.client)
+
+        # there should be 3 tests, we start watch operation and have 2 set operation on same key
+        spans = self.assert_span_count(3)
 
         self.assertEqual(spans[0].attributes.get("db.statement"), "WATCH ?")
         self.assertEqual(spans[0].kind, SpanKind.CLIENT)